+++ tests/config/prettier-entry.js
@@ -1,0 +1,4 @@
+
+import getPrettier from "./get-prettier.js"
+
+export default await getPrettier();
